Switch to recording transitions explicitly.
This simplies the internal storage. Previously, any value that needed
interpolation was stored in a tuple (old, new). This made handling
position/rotation awkward because of the pattern matching and extraction
needed. It also wouldn't scale up with the data. For example, grabbing
something is currently binary but with improved graphics would want to
be animated. That would have required a retrofit for a lot of code,
where as now the transition can be directly observed.